requirements.txt
requests>=2.31.0
tenacity>=8.2
pyyaml>=6.0
pandas>=2.0
delta-spark>=3.0
